rename types directories to model
