+++ src/app/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './register/types';
+export * from './register/model';
