+++ src/app/auth/login/model/login-response.ts
@@ -1,0 +1,3 @@
+import { Agent, DataResponse } from '@shared/model';
+
+export type LoginResponse = DataResponse<Agent>;
