--- src/app/auth/login/types/login-response.ts
@@ -1,3 +1,0 @@
-import { Agent, DataResponse } from '@shared/types';
-
-export type LoginResponse = DataResponse<Agent>;
