+++ src/app/auth/register/model/index.ts
