+++ src/app/shared/model/faction.ts
