+++ src/app/shared/model/waypoint-trait-symbol.ts
