+++ src/app/shared/model/waypoint-trait.ts
