+++ src/app/shared/model/waypoint-type.ts
