+++ src/app/shared/model/waypoint.ts
